slip_detection/requirements.txt
numpy>=1.25.2
hydra-core
pyzmq
scipy
opencv-python
gdown
gradio

--extra-index-url https://download.pytorch.org/whl/cpu
torch
